+++ src/pages/Home.jsx
@@ -1,0 +1,15 @@
+import '../styles/page_styles/Home.css';
+
+import Header from '../components/Header';
+import Main from '../components/Main'
+import Footer from '../components/Footer';
+
+export default function Home() {
+    return (
+        <>
+            <Header />
+            <Main />
+            <Footer />
+        </>
+    )
+}
