--- src/pages/home/Home.jsx
@@ -1,15 +1,0 @@
-import '../../styles/page_styles/Home.css';
-
-import Header from '../../components/header/Header';
-import Footer from '../../components/footer/Footer';
-import Main from '../../components/main/Main'
-
-export default function Home() {
-    return (
-        <>
-            <Header />
-                <Main />
-            <Footer />
-        </>
-    )
-}
