--- src/styles/page_styles/Login.css
@@ -1,78 +1,0 @@
-.login-wrapper {
-    width: 80%;
-    background-color: white;
-    margin: 0 auto;
-    padding: 100px 40px 150px 40px;
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-}
-
-.form-area {
-    width: 400px;
-}
-
-.form-area .form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.form .email,
-.form .password {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.form .email p,
-.form .password p {
-    margin: 0;
-    color: #1f1f1f;
-    font-weight: bold;
-}
-
-.form .email>input, .password>input {
-    width: 360px;
-    height: 36px;
-    padding: 8px 10px;
-    margin: 10px 0 16px 0;
-    font-size: 18px;
-    border-radius: 8px;
-    border: none;
-    outline: none;
-    color: #1f1f1f;
-    background-color: #f8f8f8
-}
-
-.form .password>input {
-    margin-bottom: 30px;
-}
-
-.form button {
-    width: 360px;
-    height: 36px;
-    border: none;
-    background: linear-gradient(to left, #fd0, #eb0);
-    color: #1f1f1f;
-    font-weight: bold;
-    font-size: 14px;
-    border-radius: 8px;
-}
-
-.login-footer {
-    width: 360px;
-    display: flex;
-    justify-content: space-between;
-    background: none;
-}
-
-.login-footer p {
-    margin: 0;
-}
-
-.login-footer .footer-link {
-    color: orange;
-    font-size: 12px;
-    text-decoration: underline;
-}
