--- src/styles/page_styles/Sobre.css
@@ -1,33 +1,0 @@
-.wrapper {
-    width: 80%;
-    background-color: white;
-    padding: 100px 40px;
-    margin: 0 auto;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-
-.title-wrapper {
-    display: flex;
-    align-items: center;
-    margin-bottom: 40px;
-}
-
-.title-wrapper .title-cursor {
-    height: 23px;
-    width: 6px;
-    margin-top: 4px;
-    background-color: rgb(255, 238, 0);
-
-    margin-right: 12px;
-}
-
-p {
-    margin-left: 17px;
-}
-
-.back-img {
-    margin: 0 17px 17px 17px;
-}
